refactor: share trade history mapping
- extract UserTradeHistoryMapper with loadListings and toOutput logic shared by both use cases
- update GetMyPurchases and GetMySales to delegate mapping to UserTradeHistoryMapper
- register UserTradeHistoryMapper bean in UseCaseConfig and inject into both use case beans
- update GetMyPurchasesTest and GetMySalesTest to build the mapper from the real class
