+++ src/main/java/br/com/eightbitbazar/application/usecase/user/GetMySales.java
@@ -1,29 +1,21 @@
 package br.com.eightbitbazar.application.usecase.user;
 
 import br.com.eightbitbazar.application.port.in.GetMySalesUseCase;
-import br.com.eightbitbazar.application.port.out.ListingRepository;
 import br.com.eightbitbazar.application.port.out.PurchaseRepository;
-import br.com.eightbitbazar.domain.listing.Listing;
-import br.com.eightbitbazar.domain.listing.ListingId;
 import br.com.eightbitbazar.domain.purchase.Purchase;
 import br.com.eightbitbazar.domain.user.UserId;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 
-import java.util.Map;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
 public class GetMySales implements GetMySalesUseCase {
 
     private final PurchaseRepository purchaseRepository;
-    private final ListingRepository listingRepository;
+    private final UserTradeHistoryMapper tradeHistoryMapper;
 
-    public GetMySales(PurchaseRepository purchaseRepository, ListingRepository listingRepository) {
+    public GetMySales(PurchaseRepository purchaseRepository, UserTradeHistoryMapper tradeHistoryMapper) {
         this.purchaseRepository = purchaseRepository;
-        this.listingRepository = listingRepository;
+        this.tradeHistoryMapper = tradeHistoryMapper;
     }
 
     @Override
@@ -32,32 +24,7 @@ public class GetMySales implements GetMySalesUseCase {
             userId,
             PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "createdAt"))
         );
-        Map<ListingId, Listing> listingsById = loadListings(purchases);
-
-        return purchases.map(purchase -> toOutput(purchase, listingsById.get(purchase.listingId())));
-    }
-
-    private Map<ListingId, Listing> loadListings(Page<Purchase> purchases) {
-        Set<ListingId> listingIds = purchases.getContent().stream()
-            .map(Purchase::listingId)
-            .collect(Collectors.toSet());
-
-        return listingRepository.findByIds(listingIds).stream()
-            .collect(Collectors.toMap(Listing::id, Function.identity()));
-    }
-
-    private UserTradeHistoryItemOutput toOutput(Purchase purchase, Listing listing) {
-        return new UserTradeHistoryItemOutput(
-            purchase.id(),
-            purchase.listingId().value(),
-            listing != null ? listing.name() : null,
-            listing != null ? listing.type().name() : null,
-            listing != null ? listing.status().name() : null,
-            purchase.amount(),
-            purchase.finalAmount(),
-            purchase.paymentMethod() != null ? purchase.paymentMethod().name() : null,
-            purchase.status().name(),
-            purchase.createdAt()
-        );
+        return tradeHistoryMapper.map(purchases);
     }
 }
+
